+++ src/app/components/Header/header.js
@@ -65,7 +65,7 @@ class Header extends React.Component {
             notifySuccess('avatar changed successfully!');
         }
         catch(error) {
-            if(error.response.status == 401) {
+            if(error.response && error.response.status == 401) {
                 this.props.history.push('/');
                 localStorage.clear();
                 notifyError('you are not logged in');
