+++ src/app/components/Overview/overview.js
@@ -33,7 +33,7 @@ class Overview extends React.Component {
             this.setState({ notes:response.data.data.notes });
         }
         catch(error) {
-            if(error.response.status == 401) {
+            if(error.response && error.response.status == 401) {
                 this.props.history.push('/');
                 localStorage.clear();
                 notifyError('you are not logged in');
@@ -53,7 +53,7 @@ class Overview extends React.Component {
             this.setState({ activities:activities, activityPage:page, totalActivities:totalActivities, requestActive:false, fetchedData:true });
         }
         catch(error) {
-            if(error.response.status == 401 && page > 1) {
+            if(error.response && error.response.status == 401 && page > 1) {
                 this.props.history.push('/');
                 localStorage.clear();
                 notifyError('you are not logged in');
